requirements.txt
pyperclip==1.8.2
google-cloud-storage==2.10.0
python-dotenv==1.0.0
Pillow==10.0.0
pillow-heif==0.13.0
